+++ com.unity.formats.fbx/Editor/ExportModelSettings.cs
@@ -95,6 +95,15 @@ namespace UnityEditor.Formats.Fbx.Exporter
             exportSettings.SetExportUnredererd(EditorGUILayout.Toggle(exportSettings.ExportUnrendered));
             EditorGUI.EndDisabledGroup ();
             GUILayout.EndHorizontal ();
+            
+            GUILayout.BeginHorizontal();
+            EditorGUILayout.LabelField(new GUIContent("Preserve Import Settings",
+                "If checked, the import settings from the overwritten FBX will be carried over to the new version."), GUILayout.Width(LabelWidth - FieldOffset));
+            // greyed out if exporting outside assets folder
+            EditorGUI.BeginDisabledGroup(ExportSettings.instance.ExportOutsideProject);
+            exportSettings.SetPreserveImportSettings(EditorGUILayout.Toggle(exportSettings.PreserveImportSettings));
+            EditorGUI.EndDisabledGroup();
+            GUILayout.EndHorizontal();
         }
     }
 
@@ -107,6 +116,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         bool UseMayaCompatibleNames { get; }
         bool AllowSceneModification { get; }
         bool ExportUnrendered { get; }
+        bool PreserveImportSettings { get; }
         Transform AnimationSource { get; }
         Transform AnimationDest { get; }
     }
@@ -152,6 +162,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         public abstract ExportSettings.LODExportType LODExportType { get; }
         public abstract ExportSettings.ObjectPosition ObjectPosition { get; }
         public abstract bool ExportUnrendered { get; }
+        public virtual bool PreserveImportSettings { get;  }
         public abstract bool AllowSceneModification { get; }
     }
 
@@ -166,6 +177,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         private ExportSettings.ObjectPosition objectPosition = ExportSettings.ObjectPosition.LocalCentered;
         [SerializeField]
         private bool exportUnrendered = true;
+        private bool preserveImportSettings = false;
 
         public override ExportSettings.Include ModelAnimIncludeOption { get { return include; } }
         public void SetModelAnimIncludeOption(ExportSettings.Include include) { this.include = include; }
@@ -175,6 +187,8 @@ namespace UnityEditor.Formats.Fbx.Exporter
         public void SetObjectPosition(ExportSettings.ObjectPosition objPos){ this.objectPosition = objPos; }
         public override bool ExportUnrendered { get { return exportUnrendered; } }
         public void SetExportUnredererd(bool exportUnrendered){ this.exportUnrendered = exportUnrendered; }
+        public override bool PreserveImportSettings { get { return preserveImportSettings; } }
+        public void SetPreserveImportSettings(bool preserveImportSettings){ this.preserveImportSettings = preserveImportSettings && !ExportSettings.instance.ExportOutsideProject; }
         public override bool AllowSceneModification { get { return false; } }
     }
 }